requirements.txt
bcrypt==4.3.0
black==25.1.0
blinker==1.9.0
boto3==1.37.6
botocore==1.37.6
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
dill==0.3.6
dnslib==0.9.26
dnspython==2.7.0
ecdsa==0.19.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
localstack==4.2.0
localstack-core==4.2.0
localstack-ext==4.2.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
plux==1.12.1
psutil==7.0.0
pyasn1==0.4.8
pycparser==2.22
Pygments==2.19.1
pyotp==2.9.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
rsa==4.9
s3transfer==0.11.4
semver==3.0.4
six==1.17.0
SQLAlchemy==2.0.38
tabulate==0.9.0
tailer==0.4.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
